requirements.txt
Flask==3.0.3
line-bot-sdk==3.4.0
groq==0.9.0
python-dotenv==1.0.1
chromadb==0.5.23